### 2025-04-17 #7965 @rkurchin
👋 @Linux-cpp-lisp and/or @ameya98, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
