### 2025-04-23 848ad97
update supernova script
