+++ scripts/supernova.py
@@ -3,25 +3,55 @@ import os
 import time
 import argparse
 import numpy as np
-from functools import partial
-import matplotlib.pyplot as plt
 import h5py
-from typing import Optional
+import matplotlib.pyplot as plt
+from typing import Optional, Union
 
 import jax.numpy as jnp
 import jax
 import jax.random as random
-import jax.experimental.mesh_utils as mesh_utils
 from jaxtyping import ArrayLike
 import optax
 import equinox as eqx
 
-import geometricconvolutions.geometric as geom
-import geometricconvolutions.ml as ml
-import geometricconvolutions.ml_eqx as ml_eqx
-import geometricconvolutions.utils as utils
-import geometricconvolutions.data as gc_data
-import geometricconvolutions.models_eqx as models
+import ginjax.geometric as geom
+import ginjax.ml as ml
+import ginjax.utils as utils
+import ginjax.models as models
+
+
+def plot_fields(
+    images_dir: str, density_data: jax.Array, temperature_data: jax.Array, velocity_data: jax.Array
+) -> None:
+    density_data = density_data[0, :, 32]  # (59, spatial_2D)
+    temperature_data = temperature_data[0, :, 32]
+    velocity_data = velocity_data[0, :, 32]
+    nrows = 4
+    ncols = 5
+
+    fig, axes = plt.subplots(nrows, ncols, figsize=(6 * ncols, 6 * nrows))
+    for i, row in enumerate([0, 1, 2, 29]):
+        geom.GeometricImage(jnp.log10(density_data[row]), 0, 2, False).plot(
+            axes[i, 0], f"density step {row}", colorbar=True
+        )
+        geom.GeometricImage(jnp.log10(temperature_data[row]), 0, 2, False).plot(
+            axes[i, 1], f"temperature step {row}", colorbar=True
+        )
+        vec_norm = geom.norm(2 + 1, velocity_data, keepdims=True)
+        scaled_velocity_data = (jnp.log10(vec_norm) / vec_norm) * velocity_data
+        geom.GeometricImage(scaled_velocity_data[row, ..., 0], 0, 2, False).plot(
+            axes[i, 2], f"velocity_x step {row}", colorbar=True
+        )
+        geom.GeometricImage(scaled_velocity_data[row, ..., 1], 0, 2, False).plot(
+            axes[i, 3], f"velocity_y step {row}", colorbar=True
+        )
+        geom.GeometricImage(scaled_velocity_data[row, ..., 2], 0, 2, False).plot(
+            axes[i, 4], f"velocity_z step {row}", colorbar=True
+        )
+
+    plt.tight_layout()
+    plt.savefig(f"{images_dir}supernova_steps.png")
+    plt.close(fig)
 
 
 def read_one_h5(filename: str, input_step: int, output_step: int) -> tuple:
@@ -39,6 +69,9 @@ def read_one_h5(filename: str, input_step: int, output_step: int) -> tuple:
     't1_fields': ['velocity'], (16,59,64,64,64,3) presumably (batch,timesteps,spatial,tensor)
     't2_fields': empty
 
+    Note that the batch size of 16 is for the training files, valid and test only have 2. Thus:
+    There are 400 Msun_1 total training trajectories, 50 total valid traj, and 50 total test traj
+
     args:
         filename (str): the full file path
         data_class (str): either 'train', 'test', or 'valid'
@@ -48,7 +81,8 @@ def read_one_h5(filename: str, input_step: int, output_step: int) -> tuple:
 
     scalar_X_ls = []
     scalar_y_ls = []
-    for scalar_field in ["density", "pressure", "temperature"]:
+    # only use density and temperature omitting pressure as in Keiya's paper
[…]
