### 2025-04-25 f18bbdf
add covariant tensor axes support for geometric images
