+++ tests/test_geometric_image.py
@@ -1,11 +1,13 @@
 import math
 import time
 import numpy as np
+import pytest
+
+import jax.numpy as jnp
+import jax
+from jax import random
 
 import ginjax.geometric as geom
-import pytest
-import jax.numpy as jnp
-from jax import random
 
 TINY = 1.0e-5
 
@@ -51,14 +53,17 @@ class TestGeometricImage:
         image1 = geom.GeometricImage.zeros(20, 0, 0, 2)
         assert image1.data.shape == (20, 20)
         assert image1.k == 0
+        assert image1.covariant_axes == ()
 
         image2 = geom.GeometricImage.zeros(20, 1, 0, 2)
         assert image2.data.shape == (20, 20, 2)
         assert image2.k == 1
+        assert image2.covariant_axes == (False,)
 
         image3 = geom.GeometricImage.zeros(20, 3, 0, 2)
         assert image3.data.shape == (20, 20, 2, 2, 2)
         assert image3.k == 3
+        assert image3.covariant_axes == (False, False, False)
 
         # square but spatial_dims in constructor, odd parity, D=3, not torus
         image4 = geom.GeometricImage.zeros((5,) * 3, 1, 1, 3, False)
@@ -67,11 +72,13 @@ class TestGeometricImage:
         assert image4.D == 3
         assert image4.parity == 1
         assert image4.is_torus == (False,) * 3
+        assert image4.covariant_axes == (False,)
 
         # non-square, D=3
         image5 = geom.GeometricImage.zeros((4, 5, 6), 0, 0, 3)
         assert image5.data.shape == (4, 5, 6)
         assert image5.D == 3
+        assert image5.covariant_axes == ()
 
         # 1D
         image6 = geom.GeometricImage.zeros((5,), 0, 0, 1, True)
@@ -80,6 +87,7 @@ class TestGeometricImage:
         assert image6.D == 1
         assert image6.parity == 0
         assert image6.is_torus == (True,)
+        assert image6.covariant_axes == ()
 
         image7 = geom.GeometricImage.zeros((5,), 0, 1, 1, False)
         assert image7.data.shape == (5,)
@@ -87,6 +95,23 @@ class TestGeometricImage:
         assert image7.D == 1
         assert image7.parity == 1
         assert image7.is_torus == (False,)
+        assert image7.covariant_axes == ()
+
+        # covariant scalar, so still nothing
+        image8 = geom.GeometricImage.zeros(20, 0, 0, 2, covariant_axes=True)
+        assert image8.covariant_axes == ()
+
+        # covariant vector
+        image9 = geom.GeometricImage.zeros(20, 1, 0, 2, covariant_axes=True)
+        assert image9.covariant_axes == (True,)
+
+        # fully covariant 2-tensor
+        image10 = geom.GeometricImage.zeros(20, 2, 0, 2, covariant_axes=True)
+        assert image10.covariant_axes == (True, True)
+
+        # mixed 2-tensor
+        image11 = geom.GeometricImage.zeros(20, 2, 0, 2, covariant_axes=(True, False))
+        assert image11.covariant_axes == (True, False)
 
     def testConstructor(self):
         # note we are not actually relying on randomness in this function, just filling values
@@ -97,28 +122,34 @@ class TestGeometricImage:
         assert image1.D == 2
         assert image1.k == 0
         assert image1.is_torus == (True,) * 2
+        assert image1.covariant_axes == ()
 
         image2 = geom.GeometricImage(random.uniform(key, shape=(10, 10, 2)), 0, 2)
         assert image2.data.shape == (10, 10, 2)
         assert image2.D == 2
         assert image2.k == 1
+        assert image2.covariant_axes == (False,)
 
[…]
