+++ src/ginjax/geometric/functional_geometric_image.py
@@ -571,10 +571,11 @@ def times_group_element(
     parity: int,
     gg: np.ndarray,
     precision: Optional[jax.lax.Precision] = None,
+    covariant_axes: Union[bool, tuple[bool, ...]] = False,
 ) -> jax.Array:
     """
-    Apply a group element of SO(2) or SO(3) to the geometric image. First apply the action to the location of the
-    pixels, then apply the action to the pixels themselves.
+    Apply a group element of SO(2) or SO(3) to the geometric image. First apply the action to the
+    location of the pixels, then apply the action to the pixels themselves.
 
     args:
         D: dimension of the data
@@ -582,13 +583,20 @@ def times_group_element(
         parity: parity of the data, 0 for even parity, 1 for odd parity
         gg: a DxD matrix that rotates the tensor. Note that you cannot vmap
             by this argument because it needs to deal with concrete values
-        precision: eisnum precision, normally uses lower precision, use
-            jax.lax.Precision.HIGH for testing equality in unit tests
+        precision: einsum precision, normally uses lower precision, use jax.lax.Precision.HIGHEST
+            for testing equality in unit tests
+        covariant_axes: which of k tensor axes are covariant, i.e. they rotate covariantly
+            with the coordinate change. False for typical vectors, true for gradients.
 
     returns:
         the rotated image data
     """
     spatial_dims, k = parse_shape(data.shape, D)
+    if isinstance(covariant_axes, bool):
+        covariant_axes = (covariant_axes,) * k
+
+    assert len(covariant_axes) == k
+
     sign, _ = jnp.linalg.slogdet(gg)
     parity_flip = sign**parity  # if parity=1, the flip operators don't flip the tensors
 
@@ -603,7 +611,8 @@ def times_group_element(
         # vector, by the group action. The image pixels have already been rotated.
         einstr = LETTERS[: len(data.shape)] + ","
         einstr += ",".join([LETTERS[i + 13] + LETTERS[i + D] for i in range(k)])
-        tensor_inputs = (rotated_pixels,) + k * (gg,)
+        rotation_matrices = tuple(gg.T if covariant else gg for covariant in covariant_axes)
+        tensor_inputs = (rotated_pixels,) + rotation_matrices
         newdata = jnp.einsum(einstr, *tensor_inputs, precision=precision) * (parity_flip)
 
     return newdata
