+++ src/ginjax/geometric/geometric_image.py
@@ -9,7 +9,7 @@ import jax.numpy as jnp
 from jax.tree_util import register_pytree_node_class
 from jaxtyping import ArrayLike
 
-from ginjax.geometric.constants import LeviCivitaSymbol, KroneckerDeltaSymbol, TINY
+from ginjax.geometric.constants import LeviCivitaSymbol, KroneckerDeltaSymbol, TINY, LETTERS
 from ginjax.geometric.functional_geometric_image import (
     average_pool,
     convolve,
@@ -36,6 +36,7 @@ class GeometricImage:
     D: int
     spatial_dims: tuple[int, ...]
     k: int
+    covariant_axes: tuple[bool, ...]  # can be () for k==0
     data: jax.Array
     parity: int
     is_torus: tuple[bool, ...]
@@ -50,6 +51,7 @@ class GeometricImage:
         parity: int,
         D: int,
         is_torus: Union[bool, tuple[bool]] = True,
+        covariant_axes: Union[bool, tuple[bool, ...]] = False,
     ) -> Self:
         """
         Zero constructor for GeometricImage.
@@ -60,13 +62,15 @@ class GeometricImage:
             parity: 0 or 1, 0 is normal vectors, 1 is pseudovectors
             D: dimension of the image, and length of vectors or side length of matrices or tensors.
             is_torus: whether the datablock is a torus, used for convolutions
+            covariant_axes: which of k tensor axes are covariant, i.e. they rotate covariantly
+                of the coordinate change. False for typical vectors, true for gradients.
 
         returns:
             constructed GeometricImage
         """
         spatial_dims = N if isinstance(N, tuple) else (N,) * D
         assert len(spatial_dims) == D
-        return cls(jnp.zeros(spatial_dims + (D,) * k), parity, D, is_torus)
+        return cls(jnp.zeros(spatial_dims + (D,) * k), parity, D, is_torus, covariant_axes)
 
     @classmethod
     def fill(
@@ -76,6 +80,7 @@ class GeometricImage:
         D: int,
         fill: Union[jax.Array, float],
         is_torus: Union[bool, tuple[bool, ...]] = True,
+        covariant_axes: Union[bool, tuple[bool, ...]] = False,
     ) -> Self:
         """
         Fill constructor to construct a geometric image every pixel as fill
@@ -86,6 +91,8 @@ class GeometricImage:
             D: dimension of the image, and length of vectors or side length of matrices or tensors.
             fill: tensor to fill the image with
             is_torus: whether the datablock is a torus, used for convolutions. Defaults to true.
+            covariant_axes: which of k tensor axes are covariant, i.e. they rotate covariantly
+                of the coordinate change. False for typical vectors, true for gradients.
 
         returns:
             Constructed GeometricImage
@@ -101,7 +108,7 @@ class GeometricImage:
         data = jnp.stack([fill for _ in range(np.multiply.reduce(spatial_dims))]).reshape(
             spatial_dims + (D,) * k
         )
-        return cls(data, parity, D, is_torus)
+        return cls(data, parity, D, is_torus, covariant_axes)
 
     def __init__(
         self: Self,
@@ -109,6 +116,7 @@ class GeometricImage:
         parity: int,
         D: int,
         is_torus: Union[bool, tuple[bool, ...]] = True,
+        covariant_axes: Union[bool, tuple[bool, ...]] = False,
     ) -> None:
         """
         Constructor for GeometricImage. It will be (N^D x D^k), so if N=100, D=2, k=1, then it's
@@ -121,6 +129,11 @@ class GeometricImage:
             is_torus: whether the datablock is a torus, used for convolutions.
                 Takes either a tuple of bools of length D specifying whether each dimension is toroidal,
[…]

[…]
