### 2025-04-27 #7965 @rkurchin
👋 @zhanglw0521 and/or @cw-tan, would you be willing to review this submission for JOSS? We carry out our checklist-driven reviews here in GitHub issues and follow these guidelines: https://joss.readthedocs.io/en/latest/review_criteria.html
