### 2025-04-28 #8129 @editorialbot
[REVIEW]: ginjax: E(d)-Equivariant CNN for Tensor Images
**Submitting author:** @WilsonGregory (<a href="http://orcid.org/0000-0002-5511-0683">Wilson Gregory</a>)
**Repository:** https://github.com/WilsonGregory/ginjax
**Branch with paper.md** (empty if default branch):
**Version:** v0.1.5
