+++ tests/test_ml.py
@@ -42,8 +42,8 @@ class TestMachineLearning:
         assert len(X_batches) == len(Y_batches) == 5
         for X_batch, Y_batch in zip(X_batches, Y_batches):
             assert (
-                X_batch[(k, 0)].shape
-                == Y_batch[(k, 0)].shape
+                X_batch[((False,) * k, 0)].shape
+                == Y_batch[((False,) * k, 0)].shape
                 == (num_devices, batch_size, 1) + (N,) * D + (D,) * k
             )
 
@@ -70,13 +70,13 @@ class TestMachineLearning:
         assert len(X_batches) == len(Y_batches) == 4
         for X_batch, Y_batch in zip(X_batches, Y_batches):
             assert (
-                X_batch[(0, 0)].shape
-                == Y_batch[(0, 0)].shape
+                X_batch[((), 0)].shape
+                == Y_batch[((), 0)].shape
                 == (num_devices, batch_size, 1) + (N,) * D + (D,) * 0
             )
             assert (
-                X_batch[(1, 0)].shape
-                == Y_batch[(1, 0)].shape
+                X_batch[((False,), 0)].shape
+                == Y_batch[((False,), 0)].shape
                 == (num_devices, batch_size, 1) + (N,) * D + (D,) * 1
             )
 
@@ -103,13 +103,13 @@ class TestMachineLearning:
         assert len(X_batches) == len(Y_batches) == 4
         for X_batch, Y_batch in zip(X_batches, Y_batches):
             assert (
-                X_batch[(0, 0)].shape
-                == Y_batch[(0, 0)].shape
+                X_batch[((), 0)].shape
+                == Y_batch[((), 0)].shape
                 == (num_devices, batch_size, 2) + (N,) * D + (D,) * 0
             )
             assert (
-                X_batch[(1, 0)].shape
-                == Y_batch[(1, 0)].shape
+                X_batch[((False,), 0)].shape
+                == Y_batch[((False,), 0)].shape
                 == (num_devices, batch_size, 1) + (N,) * D + (D,) * 1
             )
 
@@ -131,8 +131,8 @@ class TestMachineLearning:
 
         new_input = ml.training.autoregressive_step(input1, one_step1, past_steps)
         assert jnp.allclose(
-            new_input[(0, 0)],
-            jnp.concatenate([input1[(0, 0)][future_steps:], one_step1[(0, 0)]]),
+            new_input[((), 0)],
+            jnp.concatenate([input1[((), 0)][future_steps:], one_step1[((), 0)]]),
         )
 
         # Example with scalar and vector field
@@ -148,12 +148,14 @@ class TestMachineLearning:
 
         new_input = ml.training.autoregressive_step(input2, one_step2, past_steps)
         assert jnp.allclose(
-            new_input[(0, 0)],
-            jnp.concatenate([input2[(0, 0)][future_steps:], one_step2[(0, 0)]]),
+            new_input[((), 0)],
+            jnp.concatenate([input2[((), 0)][future_steps:], one_step2[((), 0)]]),
         )
-        new_input_exp = new_input.expand(0, past_steps)[(1, 0)]  # (c,past_steps,spatial,tensor)
-        input_exp = input2.expand(0, past_steps)[(1, 0)]
-        step_exp = one_step2.expand(0, future_steps)[(1, 0)]
+        new_input_exp = new_input.expand(0, past_steps)[
+            ((False,), 0)
[…]
