+++ src/ginjax/geometric/functional_geometric_image.py
@@ -14,7 +14,7 @@ import jax.numpy as jnp
 from jaxtyping import ArrayLike
 import equinox as eqx
 
-from ginjax.geometric.constants import LETTERS
+from ginjax.geometric.constants import LETTERS, TINY
 
 
 def parse_shape(shape: tuple[int, ...], D: int) -> tuple[tuple[int, ...], int]:
@@ -589,22 +589,6 @@ def raise_lower(
     return jnp.einsum(einstr, *tensor_inputs, precision=precision)
 
 
-def get_metric_inverse(metric_tensor: jax.Array, eps: float = 1e-5) -> jax.Array:
-    # (..., D, D) -> (..., D), (..., D, D)
-    eigvals, eigvecs = jnp.linalg.eigh(metric_tensor, symmetrize_input=False)
-    D = eigvals.shape[-1]
-
-    eigvals_inv = 1.0 / (eigvals + eps)  # (...,D)
-    S_diag = jax.vmap(jnp.diag)(eigvals_inv.reshape((-1, D))).reshape(eigvals.shape + (D,))
-    # do U S U^T, and multiply each vector in centered_img by the resulting matrix
-    return jnp.einsum(
-        "...ij,...jk,...kl->...il",
-        eigvecs,
-        S_diag,
-        jnp.moveaxis(eigvecs, -1, -2),
-    )
-
-
 def get_rotated_keys(D: int, data: jax.Array, gg: np.ndarray) -> np.ndarray:
     """
     Get the rotated keys of data when it will be rotated by gg. Note that we rotate the key vector indices
@@ -637,6 +621,7 @@ def times_group_element(
     data: jax.Array,
     parity: int,
     gg: np.ndarray,
+    covariant_axes: tuple[bool, ...],
     precision: Optional[jax.lax.Precision] = None,
 ) -> jax.Array:
     """
@@ -647,8 +632,9 @@ def times_group_element(
         D: dimension of the data
         data: data block of image data to rotate, shape (spatial,tensor)
         parity: parity of the data, 0 for even parity, 1 for odd parity
-        gg: a DxD matrix that rotates the tensor. Note that you cannot vmap
-            by this argument because it needs to deal with concrete values
+        gg: a DxD matrix that rotates the tensor. Note that you cannot vmap by this argument
+            because it needs to deal with concrete values
+        covariant_axes: which axes of the tensor are covariant (True) or contravariant (False)
         precision: einsum precision, normally uses lower precision, use jax.lax.Precision.HIGHEST
             for testing equality in unit tests
 
@@ -656,6 +642,10 @@ def times_group_element(
         the rotated image data
     """
     spatial_dims, k = parse_shape(data.shape, D)
+    assert len(covariant_axes) == k, f"times_group_element: Must be {k} axes, got {covariant_axes}"
+    # hope to relax this constraint in the future, for now this function is actually unchanged
+    # because gg.inv == gg.T
+    assert jnp.allclose(gg.T @ gg, jnp.eye(D)), "times_group_element: must be subgroup of O(d)"
     sign, _ = jnp.linalg.slogdet(gg)
     parity_flip = sign**parity  # if parity=1, the flip operators don't flip the tensors
 
@@ -668,9 +658,15 @@ def times_group_element(
     else:
         # applying the rotation to tensors is essentially multiplying each index, which we can think of as a
         # vector, by the group action. The image pixels have already been rotated.
-        einstr = LETTERS[: len(data.shape)] + ","
-        einstr += ",".join([LETTERS[i + 13] + LETTERS[i + D] for i in range(k)])
-        tensor_inputs = (rotated_pixels,) + (gg,) * k
[…]
