### 2025-05-02 c3a3aba
update scripts
