### 2025-05-05 53103ac
update notebooks, docs
