### 2025-05-13 82b19de
fix some bugs, add toroidal padding for non-equivariant models
