+++ scripts/exploratory/gradient_sphere.py
@@ -2,7 +2,7 @@ import time
 import argparse
 import functools
 import numpy as np
-from typing_extensions import Any, Callable, Optional, Union
+from typing_extensions import Callable, Optional
 
 import jax.numpy as jnp
 import jax
@@ -17,72 +17,106 @@ import ginjax.models as models
 import ginjax.utils as utils
 
 
-def f(coord: jax.Array, coeffs: jax.Array) -> jax.Array:
-    x, y = coord
-    # third_degree_poly = jnp.array([1, x, y, x**2, x * y, y**2, x**3, x**2 * y, x * y**2, y**3])
-    trig = jnp.array([jnp.sin(x), jnp.sin(y), jnp.cos(x), jnp.cos(y)])
-    trig_squared = jnp.concatenate(
-        [
-            trig**2,
-            jnp.array(
-                [
-                    jnp.sin(x) * jnp.cos(x),
-                    jnp.sin(y) * jnp.cos(y),
-                    jnp.sin(x) * jnp.cos(y),
-                    jnp.sin(y) * jnp.cos(x),
-                ]
-            ),
-        ]
-    )
-    library = jnp.concatenate([trig, trig_squared])  # 12 total
-    # mixed = jnp.einsum("i,j->ij", third_degree_poly, jnp.concatenate([trig, trig_squared]))
-    # library = jnp.concatenate([third_degree_poly, mixed.reshape(-1)])
+def f(p: jax.Array, coeffs: jax.Array) -> jax.Array:
+    """
+    Scalar function on the 2d torus at point p
+
+    args:
+        p: 2d array of x,y coordinates
+        coeffs: 2*D + (2*D choose 2) (=14 for D=2) coefficients to construct our function with
+    """
+    trig = jnp.concatenate([jnp.sin(p), jnp.cos(p)])
+    trig_squared = jnp.einsum("i,j->ij", trig, trig)[jnp.triu_indices(len(trig))]
+    library = jnp.concatenate([trig, trig_squared])  # 14 total
 
     return jnp.einsum("i,i->", library, coeffs)
 
 
-def metric_f(coords: jax.Array) -> jax.Array:
-    x, y = coords
+def metric_f(p: jax.Array) -> jax.Array:
+    """
+    Metric tensor at point p of the typical round metric on a sphere.
+
+    args:
+        p: the point on the surface
+
+    returns:
+        2D metric tensor g_ij
+    """
+    x, _ = p
     return jnp.array([[1, 0], [0, jnp.sin(x) ** 2]])
 
 
-def metric_inv_f(coords: jax.Array) -> jax.Array:
-    x, y = coords
+def metric_inv_f(p: jax.Array) -> jax.Array:
+    """
+    Inverse metric tensor at point p of the typical round metric on a sphere.
+
+    args:
+        p: the point on the surface
+
+    returns:
+        2D inverse metric tensor g^ij
+    """
+    x, _ = p
     return jnp.array([[1, 0], [0, 1 / jnp.sin(x) ** 2]])
 
 
-def christoffel_f(coords: jax.Array) -> jax.Array:
-    # new dimension is added at the end
-    metric_partial = jax.jacfwd(metric_f)(coords)
+def christoffel_f(p: jax.Array) -> jax.Array:
+    """
+    Christoffel symbols at point p on a surface with a metric tensor given by metric_f
+
+    args:
+        p: the point on the surface
+
+    returns:
+        Christoffel symbol Lambda^k_ij
+    """
+    metric_partial = jax.jacfwd(metric_f)(p)
     assert isinstance(metric_partial, jax.Array)
     metric_sum = (
[…]
