### 2025-06-01 WilsonGregory/ginjax#163 @ameya98
Add instructions for reproducing Table 1
Is there an easy way to reproduce the results of Table 1? Can you list the commands needed here (and in the README)?
Thanks!

Reference: https://github.com/openjournals/joss-reviews/issues/8129
