### 2025-06-01 WilsonGregory/ginjax#164 @ameya98
Add description of GeometricImage?
The term "geometric image" as defined in your paper (https://arxiv.org/abs/2305.12585) is not a well-known term. The description in the README right now is quite insufficient and hard to understand for someone unfamiliar with your work.
It would be very helpful from an accessibility perspective to have a notebook with some visual descriptions of what geometric images are, and how your equivariant networks operate on them. Thanks!

Reference: https://github.com/openjournals/joss-reviews/issues/8129
