### 2025-06-01 WilsonGregory/ginjax#165 @ameya98
Add description of API for training and sampling models
Related to #163, it would be very helpful to have in the README a snippet of how to define, train and perform inference with your model.
Do you plan to release model checkpoints? That would help significantly with reproducibility.
Thanks!

Reference: https://github.com/openjournals/joss-reviews/issues/8129
