### 2025-06-01 WilsonGregory/ginjax#166 @ameya98
Missing 'Statement of Need' section?
From the JOSS requirements:

> A statement of need: Does the paper have a section titled 'Statement of need' that clearly states what problems the software is designed to solve, who the target audience is, and its relation to other work?

I couldn't find this in the arxiv version of your paper as listed in the README: https://arxiv.org/abs/2305.12585.

Reference: https://github.com/openjournals/joss-reviews/issues/8129
