### 2025-06-01 WilsonGregory/ginjax#167 @ameya98
Add instructions to run tests?
I see the tests are present in this folder: https://github.com/WilsonGregory/ginjax/tree/main/tests
Could you please add instructions on how these tests should be run?

Thanks!
Reference: https://github.com/openjournals/joss-reviews/issues/8129
