### 2025-06-01 WilsonGregory/ginjax#168 @ameya98
Add external documentation of core API?
It would be very helpful to host documentation for your project (eg. with https://about.readthedocs.com/) to make it easy for other people to use your models and implementations.

Reference: https://github.com/openjournals/joss-reviews/issues/8129
