### 2025-06-02 WilsonGregory/ginjax#166 @WilsonGregory
Hi @ameya98, see the JOSS paper here: https://github.com/WilsonGregory/ginjax/blob/main/joss/paper.pdf
