### 2025-06-10 206d1ea
update installation to include pytest instructions
