+++ docs/index.md
@@ -10,12 +10,16 @@ This is the documentation for the package [https://github.com/WilsonGregory/ginj
 
 ## Installation
 
-- Install using pip: `pip install ginjax`.
-- Alternatively, you can install this repo as an editable install using pip.
-  - Clone the repository `git clone https://github.com/WilsonGregory/ginjax.git`
-  - Navigate to the ginjax directory `cd ginjax`
-  - Locally install the package `pip install -e .` (may have to use pip3 if your system has both python2 and python3 installed)
-  - In order to run JAX on a GPU, you will likely need to follow some additional steps detailed in https://github.com/google/jax#installation. You will probably need to know your CUDA version, which can be found with `nvidia-smi` and/or `nvcc --version`.
+Install using pip: `pip install ginjax`.
+
+### Developer Installation
+
+To work on this package, install the repo as an editable install by doing the following:
+- Clone the repository `git clone https://github.com/WilsonGregory/ginjax.git`
+- Navigate to the ginjax directory `cd ginjax`
+- Locally install the package `pip install -e .` (may have to use pip3 if your system has both python2 and python3 installed)
+- In order to run JAX on a GPU, you will likely need to follow some additional steps detailed in https://github.com/google/jax#installation. You will probably need to know your CUDA version, which can be found with `nvidia-smi` and/or `nvcc --version`.
+- In order to run the unit tests, install pytest `pip install pytest`. Then tests can be run by calling `pytest` from the package root directory.
 
 ## Features
 
