### 2025-06-10 WilsonGregory/ginjax#163 @WilsonGregory
We could add this, but we don't have that table or any of the results in the JOSS paper so perhaps its not necessary?
