### 2025-06-26 #8129 @rkurchin
Checking in with reviewers @ameya98 and @zhanglw0521 on review progress!
