### 2025-07-04 WilsonGregory/ginjax#173 @zhanglw0521
Comments on the full documentations
The [documentation](https://ginjax.readthedocs.io/) given was easy to read and very thorough. I managed to install the package and run most of the example codes based on the guidelines. I have the following comments:

- I suggest moving the "Math Background" section ahead of "Quick Start" or even combining it with "Home". This will clarify the purpose of the package and ensure that users see the explanations of some terms before they look into the example codes. For instance, the variable `N`, which first appears in "Quick Start", is only explained in "Math Background" but nowhere before.
- In the "Group and Invariant Filters" section, and in the package, have any groups other than the d-dimensional hyperoctahedral group $B_d$ been implemented? If not, is it planned that the package will be generalized to other groups in the future? If not, I think it should be stated somewhere clearly so that people may have a better idea of what this package is supposed to deal with.
- When trying to reproduce the examples provided in the documentation, the file
   '../data/mini_2D_CFD_Rand_M0.1_Eta0.01_Zeta0.01_periodic_128_Train.hdf5'
   seemed not to be provided. I found one without the prefix `mini` under DaRUS (link provided in the documentation), though. Is the mini one a sub-sample of the complete one?
- After converting a vanilla model to its equivariant version, can the authors comment on how the number of parameters changes, please? Intuitively, I expect a decrease in the numbers because of the additional equivariance constraints, and hence the model is smaller but allows better accuracy. If this is the case, it might be worth commenting somewhere, either in the paper or in the example section. On a similar note, however, I also see:
   ```
   Equivariant Model (126905), 0.001
   ...
   Vanilla Model (87940), 0.001
   ...
   ```
    on the documentation (example 1). Do the two numbers, 126905 and 87940, mean the number of parameters? If so, why is there an increase in the number of parameters?

- Would adding colorbars that indicate the magnitude of errors to the figures under blocks [11] and [12] of Example 1 make things clearer?
- Additionally, it was mentioned that "Since our training data set is so small, the predictions don't look great". My thought was that a smaller training data set would, in general, give a smaller training error as well, and if the predictions don't look great, it is probably that the model isn't rich enough. I understand that the error is not important at all in a tutorial example, but the sentence seemed a bit nonintuitive to me.

Reference: https://github.com/openjournals/joss-reviews/issues/8129
