### 2025-07-06 9e0504c
fix typos, emphasize target audience
