### 2025-07-06 4263cab
add more details to the readme, update citation
