+++ docs/index.md
@@ -4,7 +4,7 @@
 
 This package implements the GeometricImageNet which allows for writing general functions from geometric images to geometric images. Also, with an easy restriction to group invariant CNN filters, we can write CNNs that are equivariant to those groups for geometric images.
 
-See the paper for more details: https://arxiv.org/abs/2305.12585.
+See the paper for more details: https://royalsocietypublishing.org/doi/full/10.1098/rsta.2024.0247?af=R.
 
 This is the documentation for the package [https://github.com/WilsonGregory/ginjax](https://github.com/WilsonGregory/ginjax)
 
@@ -25,11 +25,73 @@ To work on this package, install the repo as an editable install by doing the fo
 
 ### GeometricImage
 
-The GeometricImage is the main concept of this package. We define a geometric image for dimension D, spatial dimensions, parity p, and tensor order k of either contravariant or covariant axes. To construct a geometric image, do: `image = GeometricImage(data, parity, D)`. Data is a jnp.array with the shape spatial dimensions followed by `(D,)*k)`. Images default to being defined on the torus, aka periodic boundary conditions. When working with a metric that is not the flat Euclidean metric, you can define which tensor axes are contravariant (the default) or covariant.
+The main concept of this package is an object called a `GeometricImage` which generalizes a normal image to include scalar images, vector images, or tensor images of any tensor order or parity.
+For example, suppose we have data of a 2 dimensional 3 by 3 vector image.
+We could construct the geometric image as follows: 
+```python
+D = 2 # 2-dimensional image, we could also do 3D.
+# data, jax array of shape (3,3,2)
+parity = 0 # even parity, its a vector image not a pseudovector image
+image = GeometricImage(data, parity, D)
+```
+Data is a jnp.array with the shape spatial dimensions followed by `(D,)*k)`. Images default to being defined on the torus, aka periodic boundary conditions. 
+When working with a metric that is not the flat Euclidean metric, you can define which tensor axes are contravariant (the default) or covariant.
+Geometric images of a particular shape, tensor order, and parity form a vector space so we have the usual operations of addition, subtraction, and scalar multiplication:
+```python
+added_image = image + image2 # addition
+subtracted_image = image - image2 # subtraction
+scaled_image = image * 3 # scalar multiplication
+```
+Additionally, if we have two images of the same spatial dimensions, but possibly different tensor orders or parities, we can multiply them where each pixel is the tensor product.
+```python
+image1 = GeometricImage(jnp.ones((3,3,D)), parity, D)
+image2 = GeometricImage(jnp.ones((3,3,D,D)), parity, D)
+image_product = image1 * image2 # data will be shape (3,3,2,2,2)
+```
+There are many more operations we can perform on geometric images including convolutions, contractions, rotations, norms, pooling, etc.
+This example is continued in [Quickstart](https://ginjax.readthedocs.io/en/latest/quickstart/) site.
 
 ### MultiImage
+The GeometricImage class is useful for exploring and experimenting with individual geometric images; however, in machine learning contexts we typically need multiple channels and batches of images.
+In addition, to build rotationally equivariant models, we may need to track multiple image types (scalar, vector, tensor, etc.) simultaneously. 
+The MultiImage class allows us to do all of this at once, only making the assumption that the images are all in the same dimensional space (2D or 3D) and they all have the same spatial dimensions.
[…]
