### 2025-07-07 WilsonGregory/ginjax#165 @WilsonGregory
I added some small examples to the readme for creating GeometricImages and MultiImages, see https://github.com/WilsonGregory/ginjax/pull/175. I also included links in the readme to much longer examples on the doc site that should give a really good idea of how to set up and train on these datasets (https://ginjax.readthedocs.io/en/latest/examples/scalar_example/ and https://ginjax.readthedocs.io/en/latest/examples/gradient_example/).
