### 2025-07-07 8350587
update docsite pages, notebook examples
