### 2025-07-08 WilsonGregory/ginjax#173 @zhanglw0521
> Hmm, I am unable to reproduce this issue. The mini data set is in the repo here: https://github.com/WilsonGregory/ginjax/tree/main/data. It is a mini version of that larger dataset with only 4 trajectories (22 Mb) instead of 10,000 (53 Gb). I added a note to update the relative path if necessary.

Ah, my bad. It was my routine setting that caused the problem. In my case, removing `../`, or simply going into the `examples` folder helps to resolve the issue I had. I guess there's nothing to be changed here.

> Good idea, we have added colorbars.

Would it make more sense if the colorbars were scaled more consistently? It would be clearer to me if the four figures for predictions and targets, and the two figures for differences, could have the same scaling, respectively, as that would make it easier to compare the two models directly.

As for the number of parameters, I now agree that it is a bit tricky to make such a full explanation in the documentation, and I find the current one, as an example, persuasive.
