### 2025-07-08 WilsonGregory/ginjax#173 @WilsonGregory
Your suggestion for the colorbars makes sense, here are the updated version:

and
