### 2025-07-08 a5b342a
make it so readme code can actually run
