### 2025-07-08 2c5d6c8
add back added_image
