### 2025-08-05 WilsonGregory/ginjax#164 @ameya98
Looks good!
