### 2025-08-06 #8129 @rkurchin
- please avoid country abbreviations in affiliations (i.e. spell out "United States")
- line 33-34: add spaces around emdash and also make consistent singular vs. plural (probably just change post-emdash part to "images where..." rather than "an image where..."
- 35: "then value of A" -> "then the value of A"
- 39: "the tensor contraction" -> either just remove "the" or replace with "a"
- 51: hyphenate "full-fledged"
- 57, 60: hyphenate "higher-order"
- 65: hyphenated "group-invariant"
- 66: add "with" after "follow along"
