### 2025-08-10 1c19c1e
update joss paper, bump version to 0.1.6
