+++ src/ginjax/ml/__init__.py
@@ -27,8 +27,10 @@ from .training import (
     get_batches as get_batches,
     autoregressive_map as autoregressive_map,
     autoregressive_step as autoregressive_step,
+    evaluate as evaluate,
     map_loss_in_batches as map_loss_in_batches,
     map_plus_loss_in_batches as map_plus_loss_in_batches,
+    train_step as train_step,
     train as train,
     benchmark as benchmark,
     BENCHMARK_DATA as BENCHMARK_DATA,
