### 2025-08-19 334d79d
update affiliations
