### 2025-08-19 #8129 @WilsonGregory
@rkurchin I think all the updates are in. I added all the author information, made a new release and DOI, added the DOIs for the references, and made your recommended changes to the paper.

The DOI is now: https://doi.org/10.5281/zenodo.16787395.
