### 2025-08-25 #8129 @crvernon
:wave: @WilsonGregory - this submission is in good shape.  I just need you to address the following before I accept it for publication:

In your paper:

- [ ] LINE 34:  The in-text citation "...introduced in (Gregory et al., 2025) are..." should appear as "...introduced in Gregory et al. (2025) are..."
- [ ] LINE 44:  "ginjax has two main target audiences:" I think you can remove this line altogether.
- [ ] LINE 59-60:  "in particular (Brandstetter et al., 2023)." should be cited "in particular Brandstetter et al. (2023)."

Let me know once these changes have been made and I'll accept.  Thank you.
