### 2025-08-25 37cb4c8
swap citations to in text, remove line
